Creating BufferedDigitalInput and using for boulder possession
Creating a BufferedDigitalInput class that creates a list of boolean values over a short period of time and returns the most prevalent value, to filter out noise. Also implemented this class for the boulder possession IR sensor.

This necessitates a "maintainState" call throughout the program to keep the list of values updating.