Adding Translation-Based Turn Scaling
Adding a method to scale the turn based on the magnitude of translation. If the robot is being moved forward heavily, the turning sensitivity will be reduced based on the translationMaxTurnScaling value in Calibrations.java. The number specified is the maximum factor that the turning can be reduced by, but the net reduction at any point in time is scaled based on the current magnitude of translation.